refactor: rename integration/java module to integration/jvm
Rename the src/integration/java module (and its Gradle path
:src:integration:jvm) and Kotlin package
community.flock.wirespec.integration.java -> .jvm. Update the
gradle-kotest example's dependency catalog (java-jvm -> jvm-jvm),
imports, and the kotest README reference.
