+++ settings.gradle.kts
@@ -42,7 +42,7 @@ include(
     "src:converter:openapi",
     "src:integration:avro",
     "src:integration:jackson",
-    "src:integration:java",
+    "src:integration:jvm",
     "src:integration:kotest",
     "src:integration:kotlinx-serialization",
     "src:integration:wirespec",
