+++ src/integration/jvm/build.gradle.kts
